bfix: Fix Dialogs implementation using Get's Dialog
